# src/main/resources/swagger/api.yml
openapi: 3.0.1
info:
  title: OCPI 2.2.1
  description: This API exposes endpoints to manage OCPI entities.
  termsOfService: 'https://extrawest.com'
  contact:
    name: Extrawest
    url: 'https://extrawest.com'
    email: ocpi@extrawest.com
  license:
    name: MIT License
    url: 'https://choosealicense.com/licenses/mit/'
  version: '1.0'
servers:
  - url: 'http://localhost:8080'
    description: Server URL in Development environment
  - url: 'https://extrawest-api.com'
    description: Server URL in Production environment
paths:
...
  '/api/2.2.1/hubClientInfo/{country_code}/{party_id}':
    get:
      tags:
        - client-info-controller
      operationId: getHubClientInfo
      parameters:
        - name: country_code
          in: path
          required: true
          schema:
            maximum: 2
            type: string
        - name: party_id
          in: path
          required: true
          schema:
            maximum: 3
            type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ClientInfoDTO'
    put:
      tags:
        - client-info-controller
      operationId: putHubClientInfo
      parameters:
        - name: country_code
          in: path
          required: true
          schema:
            maximum: 2
            type: string
        - name: party_id
          in: path
          required: true
          schema:
            maximum: 3
            type: string
      responses:
        '200':
          description: OK
...
components:
  schemas:
...
    ClientInfoDTO:
      required:
        - country_code
        - last_updated
        - party_id
        - role
        - status
      type: object
      properties:
        party_id:
          maxLength: 3
          minLength: 1
          pattern: '/[:ascii:]/i'
          type: string
        country_code:
          maxLength: 2
          minLength: 1
          pattern: '/[:ascii:]/i'
          type: string
        role:
          type: string
          enum:
            - CPO
            - EMSP
            - HUB
            - NAP
            - NSP
            - OTHER
            - SCSP
        status:
          type: string
          enum:
            - CONNECTED
            - OFFLINE
            - PLANNED
            - SUSPENDED
        last_updated:
          type: string
          format: date-time